requirements.txt
Django==5.0.6
django-jazzmin==3.0.0
django-schema-graph==3.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
